Add a test for LowerInvoke that doesn't use "-enable-correct-eh-support"
None of the existing tests for LowerInvoke check LowerInvoke's output,
and all but one use "-enable-correct-eh-support", which is obsolete,
so those tests will be removed when that option is removed.

To make sure LowerInvoke will still have test coverage, this adds a
test for its default mode which converts invokes to calls.

Differential Revision: http://llvm-reviews.chandlerc.com/D3124

git-svn-id: http://llvm.org/svn/llvm-project/llvm/trunk@204344 91177308-0d34-0410-b5e6-96231b3b80d8